Fixes issue #2246 - Consolidate response unit tests
